adding handlers and func
